Seperated realtedPercentageAmount, getStoreDiscountCaluculation and getTotalAmount functions to calculations folder and getItemsTableRowsWithCaluculations also seperated to invoiceTable.ts file
