+++ src/calcualtions/percentageAmount.ts
@@ -1,0 +1,9 @@
+import { PERCENTAGE_FACTOR } from "../invoice";
+
+export function realtedPercentageAmount(
+    itemPrice: number,
+    percentage: number
+  ): number {
+    return itemPrice * PERCENTAGE_FACTOR * percentage;
+  }
+
