+++ src/calcualtions/totalAmount.ts
@@ -1,0 +1,3 @@
+export function getTotalAmount(price: number, gst: number, discount: number): number {
+    return price + gst - discount;
+  }
