+++ src/invoice.ts
@@ -1,6 +1,12 @@
+import exp from "constants";
+import { realtedPercentageAmount } from "./calcualtions/percentageAmount";
+// import { getStoreDiscountCaluculation } from "./calcualtions/storecalculations";
+import { getTotalAmount } from "./calcualtions/totalAmount";
 import { CustomerDataJson, Item } from "./interfaces";
 import { getInvoiceHeader } from "./invoiceHeader";
 import { getSummaryHTML } from "./summery";
+import { getItemsTableRowsWithCaluculations } from "./invoiceTable";
+import { getStoreDiscountCaluculation } from "./calcualtions/storeCalculations";
 
 export const  PERCENTAGE_FACTOR = 0.01;
 export const GST_PERCENTAGE = { essential: 1, luxery: 5, default: 10 }; // Based on categeory the percentage will be taken
@@ -9,10 +15,9 @@ export var totalAmount = 0,
   totalItemsDiscountAmount = 0,
   totalGst = 0,
   totalItemsPrice = 0;
-export function addToTotalAmountWith(itemAmount:number){
-  totalAmount+=itemAmount;
+export function addToTotalAmountWith(amount:number){
+  totalAmount+=amount;
 }
-
 export function addToTotalItemsDiscountAmountWith(discountAmount:number){
   totalItemsDiscountAmount+=discountAmount;
 }
@@ -22,38 +27,11 @@ export function addTototalGstWith(gst:number){
 export function addTotalItemsPriceWith(itemPrice:number){
   totalItemsPrice+=itemPrice;
 }
-function getItemsTableRowsWithCaluculations(item: Item): string {
-  let itemPrice = item.qty * item.rate;
-  let itemGst = realtedPercentageAmount(
-    itemPrice,
-    GST_PERCENTAGE[item.gstCategory]
-  );
-  let itemDiscountAmount = realtedPercentageAmount(itemPrice, item.discount);
-  let itemAmount = getTotalAmount(itemPrice, itemGst, itemDiscountAmount);
-
-  totalGst += itemGst;
-  totalItemsDiscountAmount += itemDiscountAmount;
-  totalItemsPrice += itemPrice;
-  totalAmount += itemAmount;
-
-  return `<tr>
-          <td>${item.name}</td>
-          <td>${item.qty}</td>
-          <td>₹${item.rate}</td>
-          <td>${item.discount}</td>
-          <td>${itemDiscountAmount}</td>
-          <td>${GST_PERCENTAGE[item.gstCategory]}</td>
-          <td>${itemGst} </td>
-          <td>₹${itemAmount}</td>
-        </tr>
-        `;
+export function subToTotalAmountWith(amount:number){
+  totalAmount-=amount;
 }
-
-function realtedPercentageAmount(
-  itemPrice: number,
-  percentage: number
-): number {
-  return itemPrice * PERCENTAGE_FACTOR * percentage;
+export function subToTotalItemPriceWith(amount:number){
+  totalItemsPrice-=amount;
 }
 
 function getContactDetails(
@@ -71,9 +49,6 @@ function getContactDetails(
     </div>`;
 }
 
-function getTotalAmount(price: number, gst: number, discount: number): number {
-  return price + gst - discount;
-}
 
 function getAllCalulationsPeformedOnTable(customerData: CustomerDataJson) {
   let html = ``;
@@ -97,22 +72,6 @@ function getAllCalulationsPeformedOnTable(customerData: CustomerDataJson) {
   return html;
 }
 
-
-function getStoreDiscountCaluculation(storeDiscounts:[]){
-  let storeDiscount = 0;
-  storeDiscounts.forEach(
-    (item: { cost: number; discount: number }) => {
-      if (totalItemsPrice >= item.cost) storeDiscount = item.discount;
-    }
-  );
-
-  let storeDiscountAmount = realtedPercentageAmount(
-    totalItemsPrice,
-    storeDiscount
-  );
-  totalAmount -= storeDiscountAmount;
-  return({storeDiscount,storeDiscountAmount})
-}
 function getDeliveryHTML(customerData: CustomerDataJson): string {
   let html = `${getInvoiceHeader()}
 <body>
